refactor: tx-kernel: move externs and types to separate files
